modified:   foundation/Foundation3/Activity.cs 	modified:   foundation/Foundation3/Program.cs 	modified:   foundation/Foundation3/SwimmingActivity.cs

+++ foundation/Foundation3/SwimmingActivity.cs
@@ -1,7 +1,7 @@
 
 public class SwimmingActivity:Activity
 {
-    private int _laps;
+    private double _laps;
     public SwimmingActivity(string date, int minutes, int laps): base(date,minutes)
     {
         this._laps = laps;
